Considerably less skelital support for using 12 and 16 bits for data storage.
changed how maxSize is determined, as the old way looked aweful with the 
special cases for 24 and 32 bit wav files.

added a make rule build-test to make my life easier.  

WARNING: as of yet, these changes have been tested to compile, but not 
for functionality, not that it matters much as you cannot yet access the 
new parts.